requirements.txt
discord.py
kspconfig
aiohttp